@@ -1,20 +1,42 @@
 package org.example;
 
-import static org.junit.Assert.assertTrue;
+import ch.qos.logback.classic.Logger;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+import org.slf4j.LoggerFactory;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Unit test for simple App.
  */
 public class AppTest 
 {
+    private static Logger logger = (Logger) LoggerFactory.getLogger(AppTest.class);
+
     /**
      * Rigorous Test :-)
      */
     @Test
     public void shouldAnswerWithTrue()
     {
+        logger.error("error+++++++");
+        logger.debug("debug++++++");
+
         assertTrue( true );
+    }
+
+    @Test
+    void testMy(){
+        System.out.println("My test");
+    }
+
+    @DisplayName("Слово является палиндромом")
+    @ParameterizedTest
+    @ValueSource(strings = {"racecar", "radar", "able was I ere I saw elba"})
+    public void inPalindromeTest(String word){
+
     }
 }
